@@ -1,18 +1,11 @@
 import 'package:flutter_svg/flutter_svg.dart';
 
 import '/src/core/theme/theme.dart';
-import '/src/core/widgets/misc.dart';
+import '/src/core/widgets/common.dart';
 
 import '../intro_controller.dart';
 
 class IntrosPage extends StatefulWidget {
-  const IntrosPage({super.key});
-
-  @override
-  State<IntrosPage> createState() => _IntrosPageState();
-}
-
-class _IntrosPageState extends State<IntrosPage> {
   static const _titles = [
     'Welcome to Guardian Keyper',
     'Decentralized',
@@ -29,6 +22,20 @@ class _IntrosPageState extends State<IntrosPage> {
     'You can restore your seed phrase any time with the help of Guardians.'
         ' Even in case you’ve lost access to your device.',
   ];
+  static final _pictures = [
+    SvgPicture.asset('assets/images/intro_1.svg'),
+    SvgPicture.asset('assets/images/intro_2.svg'),
+    SvgPicture.asset('assets/images/intro_3.svg'),
+    SvgPicture.asset('assets/images/intro_4.svg'),
+  ];
+
+  const IntrosPage({super.key});
+
+  @override
+  State<IntrosPage> createState() => _IntrosPageState();
+}
+
+class _IntrosPageState extends State<IntrosPage> {
   int _step = 0;
 
   @override
@@ -36,9 +43,9 @@ class _IntrosPageState extends State<IntrosPage> {
         behavior: HitTestBehavior.translucent,
         onHorizontalDragEnd: (details) {
           if (details.velocity.pixelsPerSecond.dx < -5) {
-            if (_step == _titles.length - 1) {
+            if (_step == IntrosPage._titles.length - 1) {
               context.read<IntroController>().nextScreen();
-            } else if (_step < _titles.length - 1) {
+            } else if (_step < IntrosPage._titles.length - 1) {
               setState(() => _step++);
             }
           } else if (details.velocity.pixelsPerSecond.dx > 5 && _step > 0) {
@@ -49,15 +56,18 @@ class _IntrosPageState extends State<IntrosPage> {
           padding: paddingAll20,
           child: Column(
             children: [
-              Expanded(flex: 1, child: Container()),
+              Expanded(
+                flex: 1,
+                child: Container(),
+              ),
               Padding(
                 padding: paddingBottom32,
-                child: SvgPicture.asset('assets/images/intro_${_step + 1}.svg'),
+                child: IntrosPage._pictures[_step],
               ),
               Padding(
                 padding: paddingBottom12,
                 child: Text(
-                  _titles[_step],
+                  IntrosPage._titles[_step],
                   style: textStylePoppins620.copyWith(fontSize: 30),
                   textAlign: TextAlign.center,
                 ),
@@ -65,12 +75,15 @@ class _IntrosPageState extends State<IntrosPage> {
               Padding(
                 padding: paddingBottom20,
                 child: Text(
-                  _subtitles[_step],
+                  IntrosPage._subtitles[_step],
                   style: textStyleSourceSansPro416,
                   textAlign: TextAlign.center,
                 ),
               ),
-              Expanded(flex: 2, child: Container()),
+              Expanded(
+                flex: 2,
+                child: Container(),
+              ),
               Row(
                 mainAxisAlignment: MainAxisAlignment.spaceBetween,
                 children: [
@@ -78,14 +91,20 @@ class _IntrosPageState extends State<IntrosPage> {
                     onPressed: context.read<IntroController>().nextScreen,
                     child: Text('Skip', style: textStylePoppins616),
                   ),
-                  DotBar(
-                    count: 4,
-                    active: _step,
-                    activeColor: clBlue,
-                    passiveColor: clIndigo600,
+                  Row(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: [
+                      for (var i = 0; i < IntrosPage._titles.length; i++)
+                        Padding(
+                          padding: const EdgeInsets.only(right: 8),
+                          child: DotColored(
+                            color: i == _step ? clBlue : clIndigo600,
+                          ),
+                        ),
+                    ],
                   ),
                   TextButton(
-                    onPressed: () => _step == (_titles.length - 1)
+                    onPressed: () => _step == (IntrosPage._titles.length - 1)
                         ? context.read<IntroController>().nextScreen()
                         : setState(() => _step++),
                     child: Text('Next', style: textStylePoppins616),
