@@ -11,17 +11,10 @@ class RecoveryGroupGuardianController extends RecoveryGroupControllerBase {
 
   MessageModel? get qrCode => _qrCode;
 
-  set qrCode(MessageModel? qrCode) {
-    if (_qrCode == qrCode) return;
-    _qrCode = qrCode;
-    if (qrCode == null) return;
-    for (final e in (qrCode.payload as PeerAddressList).addresses) {
-      diContainer.networkService.addPeer(
-        qrCode.peerId,
-        e.address.rawAddress,
-        e.port,
-      );
-      nextScreen();
+  set qrCode(MessageModel? value) {
+    if (_qrCode != value) {
+      _qrCode = value;
+      if (_qrCode != null) nextScreen();
     }
   }
 }
