@@ -1,20 +1,9 @@
 import '/src/core/theme/theme.dart';
 import '/src/core/widgets/common.dart';
-import '/src/core/widgets/misc.dart';
 
 import '../restore_group_controller.dart';
 
 class ExplainerPage extends StatelessWidget {
-  static const _screenText =
-      'You can restore your Vault with the help of your Guardians. '
-      'Even in case you’ve lost access to your device.';
-  static const _screenList = [
-    'Ask a Guardian within your Vault to open their Guardian Keyper app',
-    'After that, the Guardian should navigate to “Shards” tab and find your Vault',
-    'Direct them to click “Change Vault’s Owner” and share QR Code with you',
-    'Scan the QR code',
-    'Repeat steps above for each Guardian of your Vault'
-  ];
   const ExplainerPage({super.key});
 
   @override
@@ -30,22 +19,22 @@ class ExplainerPage extends StatelessWidget {
             child: ListView(
               padding: paddingH20,
               children: [
-                const PageTitle(
-                  title: 'Ownership Transfer',
-                  subtitle: _screenText,
-                ),
-                // Numbered List
                 Container(
                   alignment: Alignment.center,
                   padding: paddingAll20,
                   decoration: boxDecoration,
-                  child: const NumberedListWidget(list: _screenList),
+                  child: const PageTitle(
+                    title: 'Restore your Vault with the help of Guardians',
+                    subtitle:
+                        'Ask a Guardian to find your Vault in the app, click '
+                        '“Change Vault’s Owner” and share QR code or Text code.',
+                  ),
                 ),
                 // Footer
                 Padding(
                   padding: paddingV32,
                   child: PrimaryButton(
-                    text: 'Scan the QR Code',
+                    text: 'Scan QR Code',
                     onPressed:
                         context.read<RestoreGroupController>().nextScreen,
                   ),
